move the icon-service to separate file
