+++ lib/items/tool-bar-button-view.js
@@ -1,31 +1,6 @@
 import {CompositeDisposable} from 'atom';
 import { ToolBarItem } from './tool-bar-item';
-
-
-// Convert the CSS class name of mdi icons to the JS name
-function mdiIconName(str) {
-  // https://github.com/Templarian/MaterialDesign-JS/blob/master/build.js#L5
-  let name = str.replace(/(-\w)/g, (matches) =>  matches[1].toUpperCase());
-  name = `${name[0].toUpperCase()}${name.slice(1)}`;
-  return `mdi${name}`;
-}
-
-let mdi; // mdi iconSet
-async function createMDIIcon(iconName, iconSize) {
-  if (!mdi) {
-    mdi = await import('@mdi/js')
-  }
-  const iconData = mdi[mdiIconName(iconName)];
-  const icon = document.createElement("div");
-  icon.style.verticalAlign = "bottom"
-  icon.innerHTML = `
-    <svg style='width:${iconSize}px; height:${iconSize}px' viewBox='0 0 24 24'>
-      <path  fill='currentColor', d='${iconData}'/>
-    <svg/>
-  `
-  return icon
-}
-
+import { createMDIIcon } from '../icon-service/mdi';
 
 /**
  * A button class with many options
